backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.0
jsonschema==4.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
aiohttp==3.9.1
requests==2.31.0
google-cloud-storage==2.10.0
google-cloud-tasks==2.16.1
google-generativeai==0.3.1
anthropic==0.7.1
email-validator==2.1.0
httpx==0.25.2
openpyxl==3.1.2
scikit-learn==1.3.2
numpy==1.24.3

